+++ src/results/electric_utility.jl
@@ -52,14 +52,20 @@ function add_electric_utility_results(m::JuMP.AbstractModel, p::AbstractInputs, 
     Year1UtilityEnergy = p.hours_per_timestep * sum(m[Symbol("dvGridPurchase"*_n)][ts, tier] 
         for ts in p.time_steps, tier in p.s.electric_tariff.n_energy_tiers)
     r["year_one_energy_supplied_kwh"] = round(value(Year1UtilityEnergy), digits=2)
-
-    GridToLoad = (sum(m[Symbol("dvGridPurchase"*_n)][ts, tier] for tier in p.s.electric_tariff.n_energy_tiers) 
-                  - sum(m[Symbol("dvGridToStorage"*_n)][b, ts] for b in p.s.storage.types) 
+    
+    if !isempty(p.s.storage.types.elec)
+        GridToLoad = (sum(m[Symbol("dvGridPurchase"*_n)][ts, tier] for tier in p.s.electric_tariff.n_energy_tiers) 
+                  - sum(m[Symbol("dvGridToStorage"*_n)][b, ts] for b in p.s.storage.types.elec) 
                   for ts in p.time_steps)
+        GridToBatt = (sum(m[Symbol("dvGridToStorage"*_n)][b, ts] for b in p.s.storage.types.elec) 
+                for ts in p.time_steps)
+    else
+        GridToLoad = (sum(m[Symbol("dvGridPurchase"*_n)][ts, tier] for tier in p.s.electric_tariff.n_energy_tiers) 
+                  for ts in p.time_steps)
+        GridToBatt = zeros(length(p.time_steps))
+    end
+    
     r["year_one_to_load_series_kw"] = round.(value.(GridToLoad), digits=3)
-
-    GridToBatt = (sum(m[Symbol("dvGridToStorage"*_n)][b, ts] for b in p.s.storage.types) 
-				  for ts in p.time_steps)
     r["year_one_to_battery_series_kw"] = round.(value.(GridToBatt), digits=3)
 
 	r["year_one_emissions_tCO2"] = round(value(m[:yr1_emissions_from_elec_grid_net_if_selected_lbs_CO2]/TONNES_TO_LBS), digits=2)
@@ -85,9 +91,9 @@ function add_electric_utility_results(m::JuMP.AbstractModel, p::MPCInputs, d::Di
                                                          tier in p.s.electric_tariff.n_energy_tiers)
     r["energy_supplied_kwh"] = round(value(Year1UtilityEnergy), digits=2)
 
-    if p.s.storage.size_kw[:elec] > 0
+    if p.s.storage.attr["ElectricStorage"].size_kwh > 0
         GridToBatt = @expression(m, [ts in p.time_steps], 
-            sum(m[Symbol("dvGridToStorage"*_n)][b, ts] for b in p.s.storage.types) 
+            sum(m[Symbol("dvGridToStorage"*_n)][b, ts] for b in p.s.storage.types.elec) 
 		)
         r["to_battery_series_kw"] = round.(value.(GridToBatt), digits=3).data
     else